refactor(web): one formatDate/formatTime for Arena history and Chat
utils/format gains formatTime (clock time in the browser locale), which Chat's
message list now imports from there instead of chatStreaming, and its tests move
alongside the other format tests. ArenaHistoryModal drops its private formatDate
and formatTime for the shared pair.
