+++ web/src/pages/Chat/__tests__/chatStreaming.test.ts
@@ -4,7 +4,6 @@ import { mockChatStream } from "../../../test/helpers";
 import { server } from "../../../test/mocks/server";
 import {
 	buildMessageContent,
-	formatTime,
 	getApiMessagesForModel,
 	streamModelResponse,
 } from "../chatStreaming";
@@ -21,41 +20,6 @@ const mockT = (key: string) => {
 	};
 	return map[key] ?? key;
 };
-
-describe("formatTime", () => {
-	it("formats timestamp as HH:MM", () => {
-		const ts = new Date("2024-01-15T14:30:00Z").getTime();
-		const result = formatTime(ts);
-
-		expect(result).toMatch(/\d{1,2}:\d{2}/);
-		expect(result).toContain(":");
-	});
-
-	it("formats midnight correctly", () => {
-		const ts = new Date("2024-01-15T00:00:00Z").getTime();
-		const result = formatTime(ts);
-
-		// Matches 00:00 (24h) or 12:00 AM (12h)
-		expect(result).toMatch(/00:00|12:00\s*AM/i);
-	});
-
-	it("formats noon correctly", () => {
-		const ts = new Date("2024-01-15T12:00:00Z").getTime();
-		const result = formatTime(ts);
-
-		// Matches 12:00 (24h) or 12:00 PM (12h)
-		expect(result).toMatch(/12:00/);
-	});
-
-	it("handles different timezones based on locale", () => {
-		const ts = Date.now();
-		const result = formatTime(ts);
-
-		expect(result).toBeDefined();
-		expect(typeof result).toBe("string");
-		expect(result).toContain(":");
-	});
-});
 
 describe("buildMessageContent", () => {
 	it("returns plain string for message without attachments", () => {
